+++ tests/unions/test_unions.py
@@ -20,7 +20,7 @@ def test_type_with_unions():
         x: Union[None, VariantA, VariantB]
         y: Union[str, VariantA]
 
-    mk_x, dict_x = p.type_constructor(X)
+    mk_x, serialize_x = p.type_constructor(X)
 
     x = mk_x({'x': {'variant_a': 1}, 'y': 'y'})
     assert isinstance(x.x, VariantA)
@@ -29,7 +29,7 @@ def test_type_with_unions():
     x = mk_x(data)
     assert isinstance(x.x, VariantB)
 
-    assert data == dict_x(x)
+    assert data == serialize_x(x)
 
     assert mk_x({'x': None, 'y': 'y'}) == mk_x({'y': 'y'})
     with pytest.raises(typeit.Invalid):
@@ -44,7 +44,7 @@ def test_type_with_primitive_union():
     class X(NamedTuple):
         x: Union[None, str]
 
-    mk_x, dict_x = type_constructor(X)
+    mk_x, serialize_x = type_constructor(X)
 
     x = mk_x({'x': None})
     assert x.x is None
@@ -53,7 +53,7 @@ def test_type_with_primitive_union():
     x = mk_x(data)
     assert x.x == 'test'
 
-    assert data == dict_x(x)
+    assert data == serialize_x(x)
 
 
 def test_union_primitive_match():
@@ -90,12 +90,12 @@ def test_test_union_primitive_and_compound_types():
     class X(NamedTuple):
         x: Union[str, Dict[str, Any]]
 
-    mk_x, dict_x = type_constructor(X)
-    mk_x_nonstrict, dict_x_nonstrict = type_constructor(X, overrides={flags.NON_STRICT_PRIMITIVES: 1})
+    mk_x, serialize_x = type_constructor(X)
+    mk_x_nonstrict, serialize_x_nonstrict = type_constructor(X, overrides={flags.NON_STRICT_PRIMITIVES: 1})
 
     data = {'x': {'key': 'value'}}
     x = mk_x(data)
-    assert dict_x(x) == data
+    assert serialize_x(x) == data
 
     x = mk_x_nonstrict(data)
-    assert dict_x_nonstrict(x) == data
+    assert serialize_x_nonstrict(x) == data
